pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ont-amplicon-phase"
dynamic = ["version"]
description = "Professional pipeline for ONT amplicon sequencing analysis and variant phasing"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Clinical Genomics Team", email = "support@example.com"},
]
keywords = ["bioinformatics", "genomics", "nanopore", "variant-calling", "phasing"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "pysam>=0.19.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "whatshap>=1.4",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.910",
]

[project.scripts]
ont-amplicon-phase = "ont_amplicon_phase.cli:main"

[tool.setuptools_scm]
write_to = "src/ont_amplicon_phase/_version.py"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true